перераспредение на несколько файлов
